feat(pit): create *pit* module and fix some clippy issues
