+++ src/arch/ia32/interrupts.rs
@@ -2,7 +2,7 @@ use core::arch::asm;
 
 pub mod idt;
 pub mod pic;
-pub mod pit8254;
+pub mod pit;
 
 /// Disable maskable external interrupts.
 ///
